serverless-rag/src/requirements.txt
onnxruntime==1.17.3
numpy>=1.24,<2.0
tokenizers>=0.19
psycopg2-binary>=2.9
